v2_services/agent_28_story_integration/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-docx>=1.1.0
